Monster color animation

+++ src/fheroes2/gui/interface_gamearea.h
@@ -47,7 +47,7 @@ enum level_t
     LEVEL_ALL = 0xFF
 };
 
-typedef std::map<std::pair<uint8_t, uint8_t>, Sprite> MapObjectSprite;
+typedef std::map<std::pair<int, uint8_t>, Sprite> MapObjectSprite;
 
 namespace Interface
 {
